requirements.txt
streamlit
numpy
opencv-python
pillow
matplotlib
moviepy
opencv-python-headless
